requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
redis>=5.0.0
mcp==1.26.0
hubspot-api-client==8.2.1
databricks-sql-connector==3.1.0
python-dotenv==1.0.1
pandas==2.2.0
openai>=1.0.0
gunicorn>=21.2.0
